requirements.txt
streamlit
matplotlib
astropy
requests
pillow
numpy